feature: include litellm integration
