+++ .env.example
@@ -98,6 +98,11 @@ XAI_API_KEY=
 # Maximum tokens for LLM responses (default: 4096).
 LLM_MAX_TOKENS=4096
 
+# LLM transport: unset or "sdk" for native vendor SDKs; "litellm" routes all API providers
+# through LiteLLM (anthropic, openai, bedrock, and OpenAI-compatible endpoints).
+# CLI-backed providers (codex, claude-code, etc.) are unaffected.
+# OPENSRE_LLM_TRANSPORT=litellm
+
 # Set the key for the provider you choose above.
 ANTHROPIC_API_KEY=
 ANTHROPIC_REASONING_MODEL=
