+++ core/llm/litellm/__init__.py
@@ -1,0 +1,7 @@
+"""LiteLLM-backed LLM client implementations.
+
+Import from the submodules directly:
+
+    from core.llm.litellm.clients import LiteLLMAgentClient, LiteLLMLLMClient
+    from core.llm.litellm.routing import build_litellm_agent_client
+"""
