+++ core/llm/sdk/__init__.py
@@ -1,0 +1,1 @@
+"""SDK-backed LLM client implementations (Anthropic, OpenAI, Bedrock)."""
