+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "anthropic>=0.77.1",
     "mcp>=1.27.0",
     "openai>=2.0.0",
+    "litellm>=1.90.0",
     # Data validation
     "pydantic>=2.12.5,<3",
     "pydantic-settings>=2.12.0,<3",
